+++ presto/testing/integration_tests/schemas/tpch/customer.sql
@@ -4,7 +4,7 @@ CREATE TABLE hive.tpch_test.customer (
     c_address VARCHAR NOT NULL,
     c_nationkey INTEGER NOT NULL,
     c_phone VARCHAR NOT NULL,
-    c_acctbal DECIMAL(15,2) NOT NULL,
+    c_acctbal DOUBLE NOT NULL,
     c_mktsegment VARCHAR NOT NULL,
     c_comment VARCHAR NOT NULL
 ) WITH (FORMAT = 'PARQUET', EXTERNAL_LOCATION = 'file:{file_path}')
