+++ presto/testing/integration_tests/schemas/tpch/orders.sql
@@ -2,7 +2,7 @@ CREATE TABLE hive.tpch_test.orders (
     o_orderkey BIGINT NOT NULL,
     o_custkey BIGINT NOT NULL,
     o_orderstatus VARCHAR NOT NULL,
-    o_totalprice DECIMAL(15,2) NOT NULL,
+    o_totalprice DOUBLE NOT NULL,
     o_orderdate DATE NOT NULL,
     o_orderpriority VARCHAR NOT NULL,
     o_clerk VARCHAR NOT NULL,
