+++ presto/testing/integration_tests/schemas/tpch/partsupp.sql
@@ -2,6 +2,6 @@ CREATE TABLE hive.tpch_test.partsupp (
     ps_partkey BIGINT NOT NULL,
     ps_suppkey BIGINT NOT NULL,
     ps_availqty BIGINT NOT NULL,
-    ps_supplycost DECIMAL(15,2) NOT NULL,
+    ps_supplycost DOUBLE NOT NULL,
     ps_comment VARCHAR NOT NULL
 ) WITH (FORMAT = 'PARQUET', EXTERNAL_LOCATION = 'file:{file_path}')
