+++ presto/testing/integration_tests/schemas/tpch/supplier.sql
@@ -4,6 +4,6 @@ CREATE TABLE hive.tpch_test.supplier (
     s_address VARCHAR NOT NULL,
     s_nationkey INTEGER NOT NULL,
     s_phone VARCHAR NOT NULL,
-    s_acctbal DECIMAL(15,2) NOT NULL,
+    s_acctbal DOUBLE NOT NULL,
     s_comment VARCHAR NOT NULL
 ) WITH (FORMAT = 'PARQUET', EXTERNAL_LOCATION = 'file:{file_path}')
